arrow-flight-rpc/requirements.txt
boto3==1.38.21
botocore==1.38.21
jmespath==1.0.1
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0
